cambios en la estruturacion del codigo y limpieza de componentes que se reciclaron 2do
